+++ app/Models/Account/Activity.php
@@ -6,17 +6,14 @@ use Parsedown;
 use App\Helpers\DateHelper;
 use App\Traits\Journalable;
 use App\Models\Contact\Contact;
+use App\Models\Journal\JournalEntry;
+use Illuminate\Database\Eloquent\Model;
+use App\Models\Instance\Emotion\Emotion;
 use App\Interfaces\IsJournalableInterface;
-use App\Models\ModelBindingHasher as Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use App\Http\Resources\Contact\ContactShort as ContactShortResource;
 
-/**
- * @property Account $account
- * @property Contact $contact
- * @property ActivityType $type
- */
 class Activity extends Model implements IsJournalableInterface
 {
     use Journalable;
@@ -40,7 +37,7 @@ class Activity extends Model implements IsJournalableInterface
      *
      * @var array
      */
-    protected $dates = ['date_it_happened'];
+    protected $dates = ['happened_at'];
 
     /**
      * The relations to eager load on every query.
@@ -84,6 +81,26 @@ class Activity extends Model implements IsJournalableInterface
     }
 
     /**
+     * Get all of the activities journal entries.
+     */
+    public function journalEntries()
+    {
+        return $this->morphMany(JournalEntry::class, 'journalable');
+    }
+
+    /**
+     * Get the emotion records associated with the activity.
+     *
+     * @return BelongsToMany
+     */
+    public function emotions()
+    {
+        return $this->belongsToMany(Emotion::class, 'emotion_activity', 'activity_id', 'emotion_id')
+            ->withPivot('account_id')
+            ->withTimestamps();
+    }
+
+    /**
      * Return the markdown parsed body.
      *
      * @return string|null
@@ -105,16 +122,6 @@ class Activity extends Model implements IsJournalableInterface
     public function getSummary()
     {
         return $this->summary;
-    }
-
-    /**
-     * Get the description for this activity.
-     *
-     * @return string or null
-     */
-    public function getDescription()
-    {
-        return $this->description;
     }
 
     /**
@@ -157,11 +164,11 @@ class Activity extends Model implements IsJournalableInterface
             'activity_type' => (! is_null($this->type) ? $this->type->name : null),
             'summary' => $this->summary,
             'description' => $this->description,
-            'day' => $this->date_it_happened->day,
-            'day_name' => mb_convert_case(DateHelper::getShortDay($this->date_it_happened), MB_CASE_TITLE, 'UTF-8'),
-            'month' => $this->date_it_happened->month,
-            'month_name' => mb_convert_case(DateHelper::getShortMonth($this->date_it_happened), MB_CASE_UPPER, 'UTF-8'),
-            'year' => $this->date_it_happened->year,
+            'day' => $this->happened_at->day,
+            'day_name' => mb_convert_case(DateHelper::getShortDay($this->happened_at), MB_CASE_TITLE, 'UTF-8'),
+            'month' => $this->happened_at->month,
+            'month_name' => mb_convert_case(DateHelper::getShortMonth($this->happened_at), MB_CASE_UPPER, 'UTF-8'),
+            'year' => $this->happened_at->year,
             'attendees' => $this->getContactsForAPI(),
         ];
     }