+++ app/Models/Instance/Emotion/Emotion.php
@@ -3,6 +3,7 @@
 namespace App\Models\Instance\Emotion;
 
 use App\Models\Contact\Call;
+use App\Models\Account\Activity;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
@@ -45,7 +46,7 @@ class Emotion extends Model
     }
 
     /**
-     * Get the call records associated with the call.
+     * Get the call records associated with the emotion.
      *
      * @return BelongsToMany
      */
@@ -53,6 +54,18 @@ class Emotion extends Model
     {
         return $this->belongsToMany(Call::class, 'emotion_call', 'emotion_id', 'call_id')
             ->withPivot('account_id', 'contact_id')
+            ->withTimestamps();
+    }
+
+    /**
+     * Get the activity records associated with the emotion.
+     *
+     * @return BelongsToMany
+     */
+    public function activities()
+    {
+        return $this->belongsToMany(Activity::class, 'emotion_activity', 'emotion_id', 'activity_id')
+            ->withPivot('account_id')
             ->withTimestamps();
     }
 }