+++ database/factories/ModelFactory.php
@@ -45,7 +45,7 @@ $factory->define(App\Models\Account\Activity::class, function (Faker\Generator $
         },
         'description' => $faker->sentence,
         'summary' => $faker->sentence,
-        'date_it_happened' => \App\Helpers\DateHelper::parseDateTime($faker->dateTimeThisCentury()),
+        'happened_at' => \App\Helpers\DateHelper::parseDateTime($faker->dateTimeThisCentury()),
     ];
 });